requirements.txt
streamlit~=1.44.1
pandas~=2.2.3
numpy~=2.2.4
scikit-learn~=1.6.1

requests~=2.32.3
nltk~=3.9.1
python-dotenv~=1.1.0
RapidFuzz~=3.13.0